Move OrdComparator, OrdStoredKey to `default` mod
